TG-30: Form logic modulation
